Move `OptionalNumber` to `types.py` part 1
